separate coposable UI
