+++ IBM_android_assignment/app/src/main/java/com/ibm/rides/ui/composable/ErrorUI.kt
@@ -1,0 +1,36 @@
+package com.ibm.rides.ui.composables
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ErrorUI(
+    errorMessage: String,
+    onRetryClick: () -> Unit
+) {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Column(
+            modifier = Modifier.padding(16.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Text(text = errorMessage)
+            Spacer(modifier = Modifier.height(8.dp))
+            Button(onClick = onRetryClick) {
+                Text("Retry")
+            }
+        }
+    }
+}
