+++ IBM_android_assignment/app/src/main/java/com/ibm/rides/ui/composable/SearchButton.kt
@@ -1,0 +1,19 @@
+package com.ibm.rides.ui.composables
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import com.ibm.rides.R
+
+@Composable
+fun SearchButton(onClick: () -> Unit) {
+    Button(
+        onClick = onClick,
+        modifier = Modifier.fillMaxWidth()
+    ) {
+        Text(stringResource(id = R.string.Search))
+    }
+}
