+++ IBM_android_assignment/app/src/main/java/com/ibm/rides/ui/composable/SearchInputField.kt
@@ -1,0 +1,26 @@
+package com.ibm.rides.ui.composables
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.res.stringResource
+import com.ibm.rides.R
+
+@Composable
+fun SearchInputField(value: String, onValueChange: (String) -> Unit) {
+    TextField(
+        value = value,
+        onValueChange = onValueChange,
+        label = { Text(stringResource(id = R.string.enter_number_of_items)) },
+        placeholder = { Text(stringResource(id = R.string.enter_number_of_items)) },
+        modifier = Modifier.fillMaxWidth(),
+        singleLine = true,
+        keyboardOptions = KeyboardOptions.Default.copy(
+            keyboardType = KeyboardType.Number
+        )
+    )
+}
